修复 rbf_knn 和 LPP 模块中的张量数据类型，确保与输入数据一致

+++ mlClustering/composition/rbf_knn.py
@@ -89,7 +89,7 @@ def knn_torch(data, n_neighbors=10, sigma=None):
     n = dist.shape[0]
     rbf_dist = rbf(dist, sigma)
 
-    S = torch.zeros_like(dist,device=data.device)
+    S = torch.zeros_like(dist,device=data.device,dtype=data.dtype)
     for i in range(n):
         index_ = torch.argsort(dist[i])[
             1 : 1 + n_neighbors
